test(LottoRankingTest) : 당첨 순위 test
