+++ src/test/java/lotto/domain/LottoStatisticsTest.java
@@ -9,7 +9,6 @@ import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 
-import static lotto.domain.LottoResult.*;
 import static org.assertj.core.api.Assertions.assertThat;
 
 class LottoStatisticsTest {
@@ -43,63 +42,8 @@ class LottoStatisticsTest {
     @Test
     void countBonusContainTest() {
         number = 7;
-        List<Integer> testCount = new ArrayList<>(List.of(0,1));
-        assertThat(lottoStatistics.countContain(number,winningSixNumbers,winningBonusNumber,count)).isEqualTo(testCount);
-    }
-
-    @DisplayName("순위를 알맞게 return 하는지 테스트(THREE)")
-    @Test
-    void findRanking5Test() {
-        LottoResult answer = THREE;
-        List<Integer> testCount = new ArrayList<>(List.of(3,0));
-        Lotto testLotto = new Lotto(List.of(1, 2, 3, 4, 5, 6));
-        testLotto.setLottoResult(LOSING);
-        lottoStatistics.findRanking(testCount,testLotto);
-        assertThat(lottoStatistics.findRanking(testCount,testLotto)).isEqualTo(answer);
-    }
-
-    @DisplayName("순위를 알맞게 return 하는지 테스트(FOUR)")
-    @Test
-    void findRanking4Test() {
-        LottoResult answer = FOUR;
-        List<Integer> testCount = new ArrayList<>(List.of(3,1));
-        Lotto testLotto = new Lotto(List.of(1, 2, 3, 4, 5, 6));
-        testLotto.setLottoResult(LOSING);
-        lottoStatistics.findRanking(testCount,testLotto);
-        assertThat(lottoStatistics.findRanking(testCount,testLotto)).isEqualTo(answer);
-    }
-
-    @DisplayName("순위를 알맞게 return 하는지 테스트(FIVE)")
-    @Test
-    void findRanking3Test() {
-        LottoResult answer = FIVE;
-        List<Integer> testCount = new ArrayList<>(List.of(5,0));
-        Lotto testLotto = new Lotto(List.of(1, 2, 3, 4, 5, 6));
-        testLotto.setLottoResult(LOSING);
-        lottoStatistics.findRanking(testCount,testLotto);
-        assertThat(lottoStatistics.findRanking(testCount,testLotto)).isEqualTo(answer);
-    }
-
-    @DisplayName("순위를 알맞게 return 하는지 테스트(FIVE_BONUS)")
-    @Test
-    void findRanking2Test() {
-        LottoResult answer = FIVE_BONUS;
-        List<Integer> testCount = new ArrayList<>(List.of(5,1));
-        Lotto testLotto = new Lotto(List.of(1, 2, 3, 4, 5, 6));
-        testLotto.setLottoResult(LOSING);
-        lottoStatistics.findRanking(testCount,testLotto);
-        assertThat(lottoStatistics.findRanking(testCount,testLotto)).isEqualTo(answer);
-    }
-
-    @DisplayName("순위를 알맞게 return 하는지 테스트(SIX)")
-    @Test
-    void findRanking1Test() {
-        LottoResult answer = SIX;
-        List<Integer> testCount = new ArrayList<>(List.of(6,0));
-        Lotto testLotto = new Lotto(List.of(1, 2, 3, 4, 5, 6));
-        testLotto.setLottoResult(LOSING);
-        lottoStatistics.findRanking(testCount,testLotto);
-        assertThat(lottoStatistics.findRanking(testCount,testLotto)).isEqualTo(answer);
+        List<Integer> testCount = new ArrayList<>(List.of(0, 1));
+        assertThat(lottoStatistics.countContain(number, winningSixNumbers, winningBonusNumber, count)).isEqualTo(testCount);
     }
 
 }
